requirements.txt
bitarray==2.9.3
pillow==11.1.0
progress==1.6
